pyproject.toml
[project]
name = "surveys-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "streamlit>=1.49.1",
    "snowflake-connector-python>=3.0.0",
    "snowflake-sqlalchemy>=1.4.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
]
